Løsning med funksjon som dekorerer
