+++ src/no/nav/statistikkapi/kandidatutfall/PresenterteOgFåttJobbenKandidaterLytter.kt
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.databind.JsonNode
 import io.micrometer.prometheus.PrometheusMeterRegistry
 import no.nav.helse.rapids_rivers.*
 import no.nav.statistikkapi.log
-import no.nav.statistikkapi.secureLog
+import no.nav.statistikkapi.logging.secure
 import no.nav.statistikkapi.stillinger.Stillingskategori
 import java.time.ZonedDateTime
 
@@ -65,7 +65,7 @@ class PresenterteOgFåttJobbenKandidaterLytter(
         val hovedmål = packet["inkludering.hovedmål"].asTextNullable()
         val utfall = Utfall.fraEventNamePostfix(eventNamePostfix)
 
-        secureLog.info(
+        secure(log).info(
             """
             aktørId: $aktørId
             organisasjonsnummer: $organisasjonsnummer
